backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2

# Google Drive API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# LlamaIndex and dependencies
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.3
llama-index-embeddings-openai>=0.1.2
llama-index-llms-openai>=0.1.2

# Document processing
pypdf>=4.0.0
python-docx>=1.0.1
pandas>=2.1.1
openpyxl>=3.1.2
pillow>=10.0.1
pycryptodome>=3.19.0

# Audio processing
pydub>=0.25.1
mutagen>=1.47.0

# Video processing
ffmpeg-python>=0.2.0

# Image processing
exifread>=3.0.0

# File type detection
python-magic>=0.4.27

# OpenAI
openai>=1.1.1

# Date handling
python-dateutil>=2.8.2